created test cases for oil and gasoline
