+++ src/model/item/__init__.py
@@ -1,4 +1,4 @@
-from .item import get_item, combine_items
+from .item_helper import get_item, combine_items
 
 
 __all__ = [
